// src/generated/models/service-governance-profile-model.ts
import type { Person } from './person-model';
import type { Service } from './service-model';
import type { Workspace } from './workspace-model';

export const ServiceGovernanceProfileStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type ServiceGovernanceProfileStatusKey = keyof typeof ServiceGovernanceProfileStatusKeyToLabel;

export interface ServiceGovernanceProfile {
  /**
   * @displayName Service Governance Profile ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Profile Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  profileName: string;
  /**
   * @displayName Auditor
   * @validationRule Required for create/update operations
   */
  auditor: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Backup Owner
   * @validationRule Required for create/update operations
   */
  backupOwner: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Maker Checker Required
   * @validationRule Required for create/update operations
   */
  makerCheckerRequired: boolean;
  /**
   * @displayName Next Review Date
   * @validationRule Required for create/update operations
   */
  nextReviewDate: string;
  /**
   * @displayName Publisher
   * @validationRule Required for create/update operations
   */
  publisher: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Report Author
   * @validationRule Required for create/update operations
   */
  reportAuthor: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Review Frequency Days
   * @validationRule Required for create/update operations
   */
  reviewFrequencyDays: number;
  /**
   * @displayName Service
   * @validationRule Required for create/update operations
   */
  service: Pick<Service, 'id' | 'serviceName'>;
  /**
   * @displayName Service Designer
   * @validationRule Required for create/update operations
   */
  serviceDesigner: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Service Owner
   * @validationRule Required for create/update operations
   */
  serviceOwner: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: ServiceGovernanceProfileStatusKey;
  /**
   * @displayName Workflow Designer
   * @validationRule Required for create/update operations
   */
  workflowDesigner: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _ServiceGovernanceProfile = 'ServiceGovernanceProfile' as const;